+++ src/PublicNavigation.spec.js
@@ -3,75 +3,30 @@ import { shallow } from 'enzyme';
 
 import PublicNavigation from './';
 
-jest.mock('./items.json', () => [{}, {}, {}]);
+jest.mock('./items.json', () => ({ items: [{}, {}, {}] }));
 
 describe('PublicNavigation', () => {
-  let nav;
+  let publicNavigation;
 
-  it('passes props to navigation without content', () => {
-    const props = {
-      a: 'prop',
-      another: 'prop',
-    };
-    nav = shallow(<PublicNavigation {...props} />);
+  it('passes items to navigation', () => {
+    publicNavigation = shallow(<PublicNavigation />);
 
-    expect(navigationWithoutContent().props()).toMatchObject(props);
+    expect(navigation().prop('items')).toEqual([{}, {}, {}]);
   });
 
-  it('passes reverse colors to navigation without content if should', () => {
-    nav = shallow(<PublicNavigation inverse />);
+  it('passes that it should have inverse colors to navigation by default', () => {
+    publicNavigation = shallow(<PublicNavigation />);
 
-    expect(isNavigationWithoutContentInverse()).toBe(true);
+    expect(navigation().prop('inverse')).toBe(true);
   });
 
-  it('does not pass reverse colors to navigation without content if should not', () => {
-    nav = shallow(<PublicNavigation inverse={false} />);
+  it('passes that it should not have inverse colors to navigation if specified', () => {
+    publicNavigation = shallow(<PublicNavigation inverse={false} />);
 
-    expect(isNavigationWithoutContentInverse()).toBe(false);
+    expect(navigation().prop('inverse')).toBe(false);
   });
 
-  it('passes items to navigation without content', () => {
-    nav = shallow(<PublicNavigation />);
-
-    expect(items()).toEqual([{}, {}, {}]);
-  });
-
-  it('passes logo link to navigation without content', () => {
-    nav = shallow(<PublicNavigation logoLink="https://transferwise.com/borderful" />);
-
-    expect(logoLink()).toEqual('https://transferwise.com/borderful');
-  });
-
-  it('allows toggling menu', () => {
-    nav = shallow(<PublicNavigation />);
-
-    expect(isMenuOpen()).toBe(false);
-    toggleMenuFromNavigationWithoutContent();
-    expect(isMenuOpen()).toBe(true);
-  });
-
-  function items() {
-    return navigationWithoutContent().prop('items');
-  }
-
-  function navigationWithoutContent() {
-    return nav.find('PublicNavigationWithoutContent');
-  }
-
-  function isNavigationWithoutContentInverse() {
-    return navigationWithoutContent().prop('inverse');
-  }
-
-  function logoLink() {
-    return navigationWithoutContent().prop('logoLink');
-  }
-
-  function toggleMenuFromNavigationWithoutContent() {
-    navigationWithoutContent().prop('onToggleMenu')();
-    nav.update();
-  }
-
-  function isMenuOpen() {
-    return navigationWithoutContent().prop('isMenuOpen');
+  function navigation() {
+    return publicNavigation.find('Navigation');
   }
 });