+++ src/PublicNavigationWithoutContent/index.js
@@ -1,3 +1,0 @@
-import PublicNavigationWithoutContent from './PublicNavigationWithoutContent';
-
-export default PublicNavigationWithoutContent;